Javier dev

+++ src/main/java/co/com/ebsa/ebsa_nexus/domain/entity/WorkRole.java
@@ -20,7 +20,7 @@ public class WorkRole {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true, length = 100)
     private String name;
@@ -41,7 +41,6 @@ public class WorkRole {
     private LocalDateTime updatedAt;
 
     public enum WorkRoleType {
-         intern,  // Personal interno
-        extern   // Personal externo
+        intern, extern
     }
 }
